separated out functions.cpp and cleaned up other files
