+++ CMakeLists.txt
@@ -4,34 +4,31 @@ project(MyProgram)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_EXPORT_COMPILE_COMMANDS TRUE)
 set(CMAKE_BUILD_TYPE Debug)
-option(BUILD_SHARED_LIBS "Enable building shared libraries" ON)
 
-include(FetchContent)
 include(GNUInstallDirs)
 
 set(CMAKE_INSTALL_LIBDIR "lib")
 set(CMAKE_INSTALL_INCLUDEDIR "include")
-set(FETCHCONTENT_FULLY_DISCONNECTED OFF)
-
-# include(cmake/fetch_parallelzone.cmake)
-# include(cmake/fetch_tensorwrapper.cmake)
 
 find_package(Eigen3 CONFIG REQUIRED)
-find_package(nlohmann_json CONFIG REQUIRED)
-find_package(Libint2 REQUIRED)
-set(SOURCES
-    src/main.cpp
-    src/functions.cpp
-)
+if(Eigen3_FOUND)
+    message(STATUS "Eigen3 found!")
+endif()
+find_package(Libint2 CONFIG REQUIRED)
+if(Libint2_FOUND)
+    message(STATUS "Libint2 found!")
+endif()
 
-add_executable(program ${SOURCES})
+file(GLOB_RECURSE src_files CONFIGURE_DEPENDS src/*.cpp)
+file(GLOB_RECURSE inc_files CONFIGURE_DEPENDS include/*.hpp)
+
+add_executable(program ${src_files} ${inc_files})
 target_include_directories(program PRIVATE include)
 target_link_libraries(program PRIVATE
     Eigen3::Eigen
-    nlohmann_json::nlohmann_json
     Libint2::cxx
-    # cereal::cereal parallelzone
 )
+
 install(TARGETS program
     RUNTIME DESTINATION bin
 )
