+++ include/energy.hpp
@@ -1,0 +1,14 @@
+#pragma once
+#include <Eigen/Dense>
+#include <libint2.hpp>
+#include <vector>
+
+// using tensor4d = Eigen::Tensor<double, 4>;
+using matrix2d = Eigen::MatrixXd;
+
+double
+nuclear_nuclear_repulsion_energy(std::vector<libint2::Atom> atoms);
+
+double compute_electronic_energy_expectation_value(matrix2d dens_mat,
+                                                   matrix2d T, matrix2d Vne,
+                                                   matrix2d G);
